chat-service/requirements.txt
grpcio>=1.64.0
grpcio-tools>=1.64.0
grpcio-health-checking>=1.64.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

